Support :association_pks_use_associated_table association option in association_pks plugin
This option handles cases where you are using the :right_primary_key
association option (or similar option), but you want the pks getter
method to return the primary key values from the associated model,
not the primary key values for the association itself.

This option is not enabled by default as it changes behavior, and
it is not obvious that such a behavior change is generally
desireable, plus it would make things slower.