Rename Dataset#to_hash to #as_hash, and add #to_hash as an alias, to allow undefing #to_hash to fix ruby calling it implicitly (Fixes #1375)
Modern ruby versions will implicitly call to_hash on arguments
if defined, even in cases where you wouldn't expect it.  For
example, using keyword arguments in ruby 2.0+:

  def foo(*a, **b)
  end

  foo(City.order(:id))

results in foo being called with a being [] and b being
City.order(:id).to_hash, which is unexpected and probably
undesired behavior.

Sequel's original name for this method was #hash_column, but in
July 2007, it was renamed to #to_hash.

In February 2013, ruby 2.0 was released with support for keyword
arguments and Kernel#Hash, which call to_hash implicitly.

The problem with reverting back to #hash_column is that currently
this method supports multiple columns as keys, and for model
datasets it doesn't even need to take a column.  #as_hash is
very similar to #to_hash in terms of being short and easy to
understand, so it seems like a reasonable choice.

Dataset#to_hash is pretty widely used, so deprecating it and
removing it would break a lot of code.  Since the existing of
Dataset#to_hash only causes problems in limited cases, leaving it
defined seems like the safer choice.  It took over 4 years for
the first bug report related to this, so it seems unlikely it is
affecting many users.