Deprecate query_literals extension
This extension changes the default behavior in a way that makes it
more likely to introduce SQL injections.  I don't think it's a
good idea to encourage its use by shipping it with Sequel.

While here, modify the deprecation messages for filter_having and
hash_aliases to encourage users to maintain it themselves.