Remove unnecessary usage of o regexp modifier
The o modifier only interpolates #{} the first time the regexp
is evaluated.  Most of the these regexps do not use #{} at all,
and therefore the o modifier has no effect.  The
MYSQL_DATABASE_DISCONNECT_ERRORS regexp does use #{}, but since
it is assigned to a constant, it does not matter if it is only
interpolated once or every time.