Deprecate Sequel::JDBC::Dataset#convert_types accessor
The setter method modified the dataset itself, which wouldn't work
in Sequel 5.  The getter method only returned dataset setting,
instead of also falling back to database setting.

In order to still support similar behavior, add a
with_convert_types method for returning a modified dataset with
the setting changed, and start storing the value in opts.