Drop support for ancient postgres driver in postgres adapter, now only pg and postgres-pr drivers are supported
This also drops the monkey-patching of PGconn and PGresult in the case
where postgres-pr is used, and significantly simplifies the logic.

Additionally, this changes the bytea conversion logic.  It always
adds the bytea escaper when postgres-pr is used.  When postgres-pr
is used, it does not modify the bytea unescaper, since the default
implementation in PG_TYPES is fine.

To simplify things, add a Database#unescape_bytea private method
when pg is used, and set the conversion proc to use that method.

The string escaping logic when using postgres-pr is also hard coded,
and now only works when force_standard_strings is true.
force_standard_strings is the default setting, so unless users turn
it off, the limitation shouldn't matter.  If anyone really needs it,
we can add conditional logic so it works in both cases.