Deprecate external modification of Sequel::JDBC::TypeConvertor
Previously, jdbc subadapters would modify this class to add
methods, then later reference the added methods.  This is
confusing and error-prone, even if it can result in slightly less
memory usage.

Switch to defining singleton methods the modules in each JDBC
subadapter.  Since Sequel::JDBC::TypeConvertor will be frozen
in Sequel 5, alias it to Type_Convertor, then deprecate
TypeConvertor.  Use Type_Convertor in the adapters that ship
with Sequel, until that code is removed in Sequel 5.

This should make sure that deprecation warnings are issued
on ruby 2.3+ if using an external JDBC adapter that accesses
TypeConvertor.