Deprecate some Sequel::Model constants
Most constants defined in Sequel::Model were not actually needed,
and it is probably better to inline them.  HOOKS is being kept
as multiple plugins use it, and RESTRICTED_SETTER_METHODS is being
kept because it needs to be referenced by an instance method.

Switch SETTER_METHOD_REGEXP usage to end_with?('=') since that
is probably faster and easier to read.