Refactor PostgreSQL conversion proc handling
This deprecates the use of PG_TYPES for storing conversion procs.
In Sequel 5, all setting of conversion procs will be done through the
Database instance.

Because of this, there will be no reason to reset conversion procs
and a lot of the complexity surrounding conversion procs can go
away.  There are no two public Database methods to add conversion
procs:

* add_conversion_proc
* add_named_conversion_proc

add_conversion_proc is new and takes an oid and a callable/block.
add_named_conversion_proc takes a type name, and block, finds
the oid by looking in the database, then calls
add_conversion_proc.

The following Database methods related to conversion procs will
be going away:

* reset_conversion_procs (public, already deprecated)
* add_named_conversion_procs
* conversion_procs_updated
* convert_named_procs_to_procs
* copy_conversion_procs
* get_conversion_procs

In the transition period, PG_TYPES is aliased to PG__TYPES,
and PG_TYPES is deprecated.  This is similar to how PG_NAMED_TYPES
was handled. Both PG_TYPES and PG__TYPES will be removed in Sequel 5.

This moves the conversion proc related code from
adapters/utils/pg_types to adapters/shared/postgres, and adds a
deprecation warning if adapters/utils/pg_types is loaded.