Refactor non sql option handling
Instead of a non_sql_options method that returns an array, have a
non_sql_option?  method that takes a key.  Then it is easy to
use super to handle nested behavior.  With this approach, we can
remove the entries in Dataset::NON_SQL_OPTIONS related to the
Sequel::Model and the mock adapter.