Deprecate Dataset#{<<,and,exclude_where,range,interval}, move to sequel_4_dataset_methods extension
These methods all have issues.

<<: This isn't really intuitive.  It does have the advantage that
you can chain it, but almost nobody does that.  Using insert
instead leads to clearer code.

and: This is just an alias for where, and while it works it doesn't
really make sense when called on a dataset without an existing
where clause.

exclude_where: This is just an alias for exclude.  It was previously
needed when exclude affected having, but with that no longer the
default behavior and the filter_having being deprecated as well,
there is no reason to keep this.

range/interval: These are both very rarely needed.  They are easily
to emulate via min and max, though that takes 2 queries instead of
1.