Undeprecate Dataset#<<
After reviewing many projects that use Sequel, this particular
method is used way more often than the other deprecated methods.
Part of the reason this method is used so much is that up until
2.11.0, it was the primary way recommended to insert records into
a dataset, showing up at the top of README.

I still think Dataset#insert is more descriptive and encourage
people to switch from #<< to #insert, but the benefits of
deprecating it do not appear to exceed the cost.