Deprecate Model.<<
This current deprecates to Dataset#<<, but it doesn't really
make sense on a model.  Intuition would lead you to believe this
calls create, when it actually ends up calling insert, skipping
hooks, validations, etc.

Unlike Dataset#<<, Model.<< seems to be rarely if ever used, so
deprecating it shouldn't cause issues in most applications.

No spec changes as there is already a deprecated spec for this
from the original Dataset#<< deprecation.