Fix buffer length passed on to writeToCopy (JRuby)
Up until now the length of the original input was passed on as length of
the buffer. However a call to `to_java_bytes` might return several bytes
for a single character depending on the encoding. In the end only part
of the actual data would have been sent to the PostgreSQL server either
leading to missing line endings and thus mixing several rows into one or
finding invalid bytes for certain encodings.