Add pg_extended_date_support extension, for handling infinite and BC dates/timestamps
The postgres adapter had support for infinite timestamps already.
This moves that support to the pg_extended_date_support, and loads
the extension if appropriate.

BC dates and timestamps were basically not handled correctly before.
Sequel used a negative year before, which PostgreSQL doesn't like
as it doesn't consider it valid date syntax.  PostgreSQL puts the BC
at the end of the string.  For dates, that's OK as ruby's Date.parse
can handle it.  For timestamps that doesn't work, as BC is
interpreted as a timezone in some cases. Move the BC to between the
date and the time when parsing the date in ruby to fix things.

Previously, there was no support at all for infinite or BC dates or
timestamps in the jdbc/postgresql adapter.  That's because
java.sql.date and java.sql.timestamp don't support BC dates at all.
Work around this by switching the typecasting to use the same
typecasting as done in the postgres adapter.

This has a few fixes for convert_infinite_timestamps=, including
handling true and using the correct conversion proc for 'f'.