Avoid warning if trying to rollback after a commit or rollback raises an exception in the postgres adapter
Sequel's behavior if a commit or rollback fails is to rollback, just
in case the failure of the commit or rollback didn't actually
close the transaction.  This doesn't cause any problems, but
PostgreSQL warns if you issue a rollback outside of a transaction.

Both ruby-pg and postgres-pr support a transaction_status method
for determining whether the connection is in a transaction or not,
and if it returns 0, it means the connection is outside of a
transaction.  So if the connection is outside of a transaction,
we skip the rollback.

I've been hesistent to do this in the past because I wasn't sure
what happens if the client and server get out of sync in regards to
what each thinks the transaction status is.  However, since all
transaction status changes are initated by the client, I don't
think this should cause any problems in practice.