Add PlaceholderCheckout dataset extension for MySQL
The MySQL adapter checks out a connection in order to do placeholder
escaping; this means that for queries with large numbers of placeholders
in them, a large amount of time can be spent checking out/checking in
connections while building up the SQL test.

This PR adds a dataset extension which wraps the _sql generation methods
with `db.synchronize`, so that only one connection is checked out for
the entire SQL generation process.

This should resolve jeremyevans/sequel#1319