Respect indislive and ignore indcheckxmin index attributes when using Database#indexes on PostgreSQL
indislive was added in PostgreSQL 9.3, and should be checked so that
if the index is in the process of being dropped, it should be
ignored.

Checking indcheckxmin is probably more likely to cause false
negatives than fix actual problems.  The problem with it is that
even if using indcheckxmin, most likely the xmin for the index
is from a previous transaction and current and future transactions
would be able to use it.  About the only case where this isn't true
is if you do something similar to:

1) begin transaction 1
2) begin transaction 2
3) create/alter/re index on transaction 2
4) call Database#indexes in transaction 1

Assuming that index operation in transaction 2 causes indcheckxmin
to be set, then Database#indexes should not return it.  The previous
code was designed to catch that.  However, it is more likely that
indcheckxmin is set but xmin for the index row is prior to the
current transaction transaction id and the transaction would be
able to use the index.

I looked into trying to check xmin for index usability if
indcheckxmin was set, but the PostgreSQL transaction id functions
due not appear to make this easy as they return a bigint that includes
transaction epoch information.  It may be possible to use the age(txid)
function to determine if the index is usable if indcheckxmin is set,
but as I am not sure how to do to that correctly, and it is very
unlikely that just ignoring indcheckxmin will cause problems in
production, ignoring it seems to be the best option at the current
point.  If anyone can provide the correct way to check xmin if
indcheckxmin is set, I'll definitely consider it.