Deprecate Dataset#where calls with no existing filter, no argument, and where the virtual row block returns nil (Fixes #1454)
This unfortunately wasn't caught in the Sequel 5 changes, because
there were no tests for it.  Basically:

  where{nil}

should be treated as:

  where(nil)

In cases where an argument was given, or when the dataset had an
existing filter, or when the filter was being inverted, or
when an argument was provided in addition to the block, the
block returning nil was already handled correctly.

I consider this a significant bug, but as this will change behavior
and it appears people may be relying on the current behavior, I'm
going to deprecate the current behavior for one version to give
developers a chance to fix their code.