Cleanup synchronize_sql extension
Use 2-space indent like the rest of the project.

Don't checkout a connection if raw SQL is used, as it would definitely
not be needed in that case.

Add support for integration testing with the extension using the
SEQUEL_SYNCHRONIZE_SQL environment variable.

Simplify the specs.  The use of the Sync module is to work around
a bug in ruby 1.9 because extend_datasets with a block does not
handle the super call correctly in that case.