Add datetime_parse_to_time extension, for parsing timestamp strings without offsets using DateTime.parse.to_time (Fixes #1455)
Ruby's Time.parse assumes the local timezone if no timezone
information is provided, and instead of raising an exception if
the time given is not valid in the local timezone (because it
occurs during a daylight savings time jump), it silents moves the
time an hour forward.

This attempts to fix that case by using DateTime.parse.to_time.
DateTime.parse assumes UTC if no timezone is given instead of
assuming local time.

This extension is only used if the timezone for the database
is set to :utc, and only if the string does not contain timezone
information.

To make sure this doesn't break anything, it copies the default
specs when it is not used, and only adds a single spec to test
for known bad times in a few common timezones.