Optimize Database#synchronize implementation on ruby 2.5+
Given the following code:

def a(&block)
  b(&block)
end

def c
  b{yield}
end

def b
  yield
end

c is about 4x faster than a on ruby 2.4, but about 10% slower than
a on ruby 2.5.  This is due to the changes in ruby 2.5 where
proc activation doesn't happen if accepting a block and just passing
it to another method.

This switches Sequel to use the same approach it uses for JRuby
also on ruby 2.5 (a is about 50% faster than c on modern versions
of JRuby).