Fix index_caching extension
Previous index caching was completely broken (referenced a local variable
not in scope), and tests didn't catch it as there weren't tests for
calling indexes.

No reason to dup the options hash at the start of indexes, nothing
modifies it.

Remove cache_schema check in indexes.  I'm not sure it should apply,
and if it should, it should be moved to the return super call above.

Drop overriding of freeze.  It was my mistake to recommend adding it, this
information should be modifiable at runtime, and is correctly protected
using a mutex.

Fix bin/sequel to wrap the table in SQL::Identifier when running
indexes.  Previously this didn't work on SQLite, but that was just
fixed.  Probably not necessary unless symbol splitting is turned on
by loading some file via -r or -L.

Support SEQUEL_INDEX_CACHING environment variable in the adapter and
integration tests, to check this doesn't break things at runtime. Fix
one SQLite adapter test that broke at runtime when index_caching was
used.