Don't attempt to insert a second time if insert_select runs a query that doesn't return results, which can happen when triggers are used
We need to differentiate the case where the INSERT query is sent
but doesn't return a result from the case where the INSERT query
is not sent.  The least breakage to backwards compatibility is
to return false instead of nil if the query runs, and have
Sequel::Model recognize the false value.