Protect migration file loading with a mutex to not break when multiple threads load migration files simultaneously
I never considered people would actually do this, but it's possible
that migrators could be run on different databases at the same time
in different threads, and it's nice to have that work correctly.

This removes the private remove_migration_classes method, no longer
needed.  It changes the behavior for load_migration_file to return
the migration class/object loaded from the migration file, and not
just populate Migration.descendants.  This is backwards
incompatible for any code that accesses Migration.descendants
expecting it to contain the migration classes, but hopefully
nobody is doing that.

This modifies the specs slightly to use Database instance variables
instead of class constants to store tables and migration files,
which is necessary for the multithreading to work.