Fix array_type for pg_array_to_many and many_to_pg_array associations in pg_array_associations plugin
The array type should be the scalar type for the array, not the
actual array type.  Previously, some of the code ended up like
ARRAY[1]::integer[][], but that didn't break anything because
integer[][] is considered the same type as integer[].

As the default type is integer and not integer[], it's best to
make sure the value is consistent. Additionally, the array_type
may be used to infer the type of the primary key that the foreign
keys reference, and in that case the scalar type must be used.