Add integer64 extension for treating Integer as a 64-bit integer when used as a generic type
Among other things, this makes primary_key and foreign_key use
64-bit integers instead of 32-bit integers by default on most
databases.

You can use the SEQUEL_INTEGER64 environment variable to load this
extension automatically when running the adapter or integration
specs.

To make PostgreSQL handle this, force a cast for the bitwise shift
test to make sure it works, as bigint << bigint is not supported.
Make a small change to a spec that checks for specific oid values.
Also, base pg_array_associations tests use an array type that
matches the primary key type used.