Treat prepared statement errors due to changing types as disconnect errors (Fixes #1481)
While these are not technically disconnect errors, treating them
as disconnect errors makes sense because it is likely to decrease
the total number of exceptions that occur if the type of a returned
column (or the number of columns) changes at runtime.

In general, if you want to make schema changes and use prepared
statements, it's best to make sure your prepared statements never
use a wildcard select, and that none of your schema changes change
the type of a column or expression returned by a prepared statement.