Fix Database#server_version when connecting to PostgreSQL 10.1+ in certain cases
Unless the connection object supported a server_version method,
Sequel called the version function, which returned a string, and
attempted to parse the information from that string using a regexp.

That's not a great way to operate, but it did work fairly well
until PostgreSQL changed their versioning in 10.  However, starting
in version 10, minor version is treated as tiny version when using
the numeric version so 10.2 should be 100002 instead of 100200.

When originally written, parsing information from the version
function was probably the best way, but in PostgreSQL 8.2 the
server_version_num setting was added.  Switch to using that
unconditionally instead of trying the server_version method on
the connection, for consistent behavior across adapters.

Update related identity code guards to use 100002 instead of
100200.