Add Database#convert_serial_to_identity on PostgreSQL 10.2+, which requires superuser access
This allows the conversion of serial columns to identity columns.
Unfortunately, PostgreSQL does not offer a way to convert an
existing serial column to an identity column.  ALTER COLUMN DROP
DEFAULT followed by ALTER COLUMN GENERATED BY DEFAULT AS IDENTITY
results in a new sequence being created, not the existing sequence.

This uses the approach provided by 2nd Quadrant, which modifies the
system tables to accomplish the change.  However, modifying the
system tables requires superuser access, so this method is limited
to superusers.  That definitely limits the cases in which the method
can be used, but I think it's still worthwhile to offer the method.

Maybe in the future PostgreSQL will add a way to convert serial
columns to identity columns without superuser access, and if so
this code can be switched to use that if it is supported by the
server.