Run Dataset#with_sql_{all,each,first,single_value} using a cached dataset to avoid clobbering the dataset's columns
Using these methods was previously risky unless you were sure
the columns used matched the dataset columns.  The documentation
made this clear, but it's still something that could bite someone.
With the changes made in Sequel 5 to freeze datasets by defaults,
the performance benefits of the with_sql_* methods can still be
had while avoiding the problems, as long as the receiver is not
an ephemeral dataset.  In most cases where the with_sql_* methods
are used for performance reasons, the receiver is not ephemeral,
so this change should hopefully not decrease performance
significantly in most cases.