Support Database#foreign_key_list :reverse option on PostgreSQL for parsing foreign key constraints that reference a given table
Database#foreign_key_list by default returns foreign keys in the
current table that reference other tables.  The :reverse option
modifies things so that it returns foreign keys in other tables
that reference the current table

When using the :reverse option, the resulting hashes contain
mostly the same data, except the :table option is table containing
the foreign key constraint instead of the table referenced by the
foreign key constraint.  :columns are still the referencing columns
(now in the other table), and :key the referenced columns (now in
the current table).

While here, unless the schema support is turned on and switches
:table to return a qualified identifier, return the :schema entry
as a separate entry in the hash.