Add pg_auto_constraint_validations plugin, for automatically converting constraint violations to validation failures on PostgreSQL
This can convert most constraint violations to validation errors.
It is intended as a last resort, as the validation error messages
it uses are pretty basic, but it can catch things that can only be
caught at INSERT/UPDATE time, as well as provide the only automatic
way to handle validation failures for CHECK and FOREIGN KEY
constraints as well as partial unique indexes.