Fix thread safety issues in connection_expiration_random_delay support
The access to @connection_expiration_random_delays wasn't synchronized
previously.  To avoid two separate syncs, store an array with the
timer and the timeout.

Additionally, use rand * delay instead of rand(delay).  rand(delay)
is not what the desired behavior is.  rand(0) is a random float
between 0 and 1, rand(1) is always 0, and rand(N) is a integer
between 0 and N-1.