Support :search_path as a shard option on PostgreSQL
Previously, :search_path was only supported as a Database option,
usage as a shard option would be ignored.

To make this possible, connection_configuration_sqls now takes
an optional opts argument for the options to use.  This defaults
to the Database opts so that any external adapters that use the
shared PostgreSQL support will not break.

To make this work on JDBC, a setup_connection_with_opts method
has been added.  This just calls setup_connection without the
opts, so it should be backwards compatible.  The jdbc/postgresql
now uses this to pass the appropriate opts to
connection_configuration_sqls.

This approach can be extended to other adapters where it makes
sense to do so, but that will probably be done on a per-request
basis.