Fix duplicate objects in nested associations when eager_graphing cascaded many_to_one=>one_to_many associations
In general the eager loaders use an identity map that is solely
based on the key, but that can cause problems with duplicate
nested associated objects if there are dependent associations that
are *_many.

When there is an actual cartesian product, the general duplicate
removal code takes care of the issue.  However, if the duplicate
removal code isn't used, then we need to have the identity map
also include a reference to the currently loading object so that
the association caches will not be shared.