Support use of LIKE with ANY function on PostgreSQL by avoiding unnecessary use of ESCAPE syntax
This allows code such as:

  Sequel.like(:column, Sequel.function(:ANY, ds))

to work correctly.  PostgreSQL supports ESCAPE for LIKE, but
does not consider LIKE ANY(subquery) ESCAPE '\\' valid syntax.

From a review of the documentation, PostgreSQL defaults to using
backslash as the LIKE escape character, so the use of ESCAPE '\\'
is unnecessary and can be dropped.

Add a feature method to determine whether ESCAPE should be used.
Turn it off for PostgreSQL, and also off for Access, and remove
the Access-specific override of the LIKE operator.