Use DB instead of Sequel::Model.db in testing doc (Fixes #1511)
Most of the rest of the documentation uses DB to represent the
Sequel::Database in use.  While Sequel::Model.db will work
in some code, it will break in other cases, such as when
using core Sequel and not Sequel::Model.  Better to be consistent,
and have people adjust the code if they aren't using the
recommended DB constant.