Fix Sequel::SQL::Window RDoc links
It seems that `{SomeClass}` isn't the right way to link classes in RDoc,
because it doesn't work (probably mixed it up with YARD). So I changed
it to the full class name, and that works. In that case the link text
will be the full constant name (`Sequel::SQL::Window`).