Make many_to_pg_array associations in the pg_array_associations plugin work on PostgreSQL 11
PostgreSQL 11 started raising a "set-valued function called in
context that cannot accept a set" error for the previous code.
Set-returning functions in the from list are kind of a smell
anyway, so even if this is a regression in PostgreSQL that will
be fixed before the final release, I think it's a good idea to
change the implementation.

This uses two different approaches for the two cases where unnest
was used for many_to_pg_array associations.  In the first case,
it does a CROSS JOIN to unnest (a lateral function call) to get
the results.  In the second case, it doesn't use unnest at all,
and switches to using EXISTS with a subquery.  The reason for the
different cases is the EXISTS approach was not easy to implement
in the first case as the the code expects to return a dataset
and not an expression.