Allow the use of IN/NOT IN operators with set returning functions for Sequel::Model datasets
The previous code assumed the right hand side of an IN/NOT IN
operator was a dataset or an array.  While those are the cases
where Sequel will automatically turn conditions specifiers into
IN/NOT IN operators, it is possible to construct operator
instances manually and such instances should still work.