Fix parsing of timetz types in the jdbc/postgresql adapter
While the timetz type is fairly braindead and nobody should use it.
The PostgreSQL JDBC doesn't handle it correctly.  It treats the
timetz type as TIME and not TIME_WITH_TIMEZONE, but doesn't expose
the timestamp used for parsing, leading to invalid values when
wrapping it in an SQLTime object.

This change just switches to using the ruby parsing, which handles
things correctly. Unfortunately, it may slow down parsing of
time types, but since few people use either time or timetz, the
loss of performance shouldn't matter much.