Refactor how emulated prepared statements are processsed
Previously, normal dataset literalization occurred on every call,
with the exception that placeholder symbols were substituted with
values.  With this change, a placeholder literalizer is used at
statement preparation, to create an array of SQL fragments, and
that array is used to create a placeholder literal string later
with the values.

This change should make prepared statement calls faster in the
emulated case.  One other advantage is that the prepared_args
array is now frozen in all cases, avoiding or at least raising
an error in cases where it would be modified later.

Other than an addition of a new spec that comes from an
existing integration test, the only spec modifications are to
force an append-only SQL string creation, as that is required
by the placeholder literalizer code, and slight tweaks to make
sure things work correctly now that prepared_sql is now calculated
during Dataset#prepare in all cases.

To make the subselect code simpler, introduce a subselect_sql_dataset
method that can be overridden separately from subselect_sql_append.