Support using Dataset#eager_graph in eager load callback for associations using join tables
While this wasn't documented, cascading eager associations to
eager_graph associations using the eager load callback worked
if the associations being eagerly loaded did not use join
tables.  Example:

  Band.eager(:albums=>proc{|ds| ds.eager_graph(:tracks)})

However, if the associations being eagerly loaded used join tables,
it didn't work.  Such associations use a subselect when graphing, and
the columns that need to be added for the eager load to work would
need to be selected inside the subselect, not outside, which means
the expressions would need to be added before calling the callback
instead of after.

That change cannot be made universally because it would break the
common case where Dataset#select is used in the callback to limit
the returned columns.  This code detects the case where eager_graph
is used in the eager_load callback, and if so, calls the callback
again with the associated columns already added.  Because datasets
are frozen, this should be safe, though it does require the callback
is idempotent.

This cannot handle cases that use both Dataset#select and
Dataset#eager_graph in the callback, but that is probably rare
enough not to worry about.