Remove Dataset#prepared_arg?
This was a private method, and is no longer called after the recent
prepared statement refactoring.  External adapters that define the
method do not need changes, but any external adapters that call the
method need to be updated to remove the call.

While here, update the code to use =~ on the symbol instead of
match on the string, should be slightly faster and simpler.