Cache eager_graph loader to speed up subsequent loads from the same dataset
This refactors the EagerGraphLoader to split initialization into
two parts: initialize and load.  initialize now returns a frozen
object, which must be dup to return a working loader.  load now
does the per-load initialization by setting up the record maps,
and freezes the loader before returning the records.

This also refactors the related dataset code so that plugins and
extensions have the ability to access the rows, object graph, and
EagerGraphLoader instance used to perform the conversion.