Do not push down expression inversion in cases where it may result in incorrect behavior (e.g. ANY/SOME/ALL operators) (Fixes #1549)
Inverting:

  (x = ANY(1))

to:

  (x != ANY(1))

is not correct.  The correct inversion is:

  NOT (x = ANY(1))

I think ANY/SOME/ALL are the only cases where this is true, but
there may be others.  Since NOT should correctly invert everything,
and inversion push down is only done to make the queries look nicer,
just use NOT for boolean 2-arity operators that are not IN/NOT IN,
where the right hand side is a SQL::Function, LiteralString, or
SQL::PlaceholderLiteralString.