Add Model#skip_validation_on_next_save!
This handles cases where save is called later in a way where you
cannot pass the :validate=>false option. It only handles the next
save call, and not future save calls. It has preference over an
explicit :validate=>true option passed to save, and as such should
be used with care (hence the !).