Do not remove currently cached many_to_one associated objects when changing the related foreign key value from nil to non-nil
If the associations cache has an entry for a many_to_one associated
object while the current foreign key value is nil, that should only
have happened if the associations cache was modified manually, in
which case it's probably best to not delete the value.  This still
deletes the value if there was a cached nil value, to avoid
caching false negatives.