Add Database#rollback_on_exit to rollback transactions instead of committing them when exiting the transaction block
Previously, all non-exception exits from a transaction block
committed the transaction.  However, there are cases where you
want to rollback the transaction but also perform a non-local
exit (such as returning out of the block or using throw).  This
method allows you to call rollback_on_exit to specify the
transaction/savepoint should be rolled back, and then perform the
non-local exit.  Previously, you could set a flag variable, then
raise Sequel::Rollback, then check the flag variable after the
transaction/savepoint block, but that usually resulted in code
that was harder to follow.

To implement this, the :savepoints entry in the internal
transaction state hash has been changed from a boolean auto_savepoint
flag to a hash.  Any external adapter or extension code that operates
on the :savepoints entry in the internal transaction state hash
will need to be modified.