Remove loading of mysqlplus in the mysql adapter
mysqlplus hasn't had a release since 2009, and mysql last had a
release in 2013.  I'm not sure to what extent mysqlplus would be
helpful anymore, but I haven't tested using it in many years,
and it's probably not a good idea to keep supporting it. It can
be added back if anyone actually reports they are using it.