More robust handling of reciprocals in one_to_one association setter when using set_associated_object_if_same
If the set_associated_object_if_same flag is set (and only the the
nested_attributes plugin currently sets it), this only clears the
currently associated object's reciprocal association if the reciprocal
object doesn't match the current object. It only sets the current
object as the newly associated object's reciprocal if the current
object is not already set as such.