Switch from using instance_exec to define_method for model associations and in some plugins
instance_exec is slow because it creates a singleton class for
each instance.  Defining a method and calling that method is
faster, sometimes by a significant amount.

This borrows the approach used in Roda, though it limits it
to only blocks that accept 0 or 1 arguments, and uses deprecation
warnings by default for arity mismatches.

This affects the associations support, as well as the
association_pks, composition, hook_class_methods, and
validation_class_methods plugins.