Make the server_block extension respect the :servers_hash Database option
Previously, the server_block extension did not respect the :servers_hash
option.  While behavior is now consistent with Sequel's default
Dataset#server behavior, this change is kind of a mixed bag, since
previously, passing an invalid server to with_server would raise an
exception (NoMethodError), and now it will not.  However, I think having
consistency is worth way to handle the situation.  If we decide to
change the default behavior to make Sequel more strict, it will benefit
both cases.